SBERDOMA-47 review fixes
