+++ apps/condo/domains/common/constants/locale.js
@@ -1,0 +1,11 @@
+const RU = require('date-fns/locale/ru')
+const EN = require('date-fns/locale/en-US')
+
+const LOCALES = {
+    ru: RU,
+    en: EN,
+}
+
+module.exports = {
+    LOCALES,
+}
