+++ apps/condo/domains/common/constants/style.js
@@ -1,5 +1,4 @@
-const { CANCELED, COMPLETED, DEFERRED, NEW_OR_REOPEND, PROCESSING } = require('./statusTypes')
-const { generate, grey, green, orange, red, yellow  } = require('@ant-design/colors')
+const { generate, grey, green  } = require('@ant-design/colors')
 
 const sberBlue = '#5EB1FC'
 const sberGreen = '#4CD174'
@@ -44,31 +43,7 @@ const antGlobalVariables = {
     '@table-header-bg': ultraLightGrey,
 }
 
-const STATUS_SELECT_COLORS = {
-    [NEW_OR_REOPEND]: {
-        color: yellow[9],
-        backgroundColor: yellow[2],
-    },
-    [PROCESSING]: {
-        color: orange[8],
-        backgroundColor: orange[4],
-    },
-    [CANCELED]: {
-        color: grey[9],
-        backgroundColor: colors.lightGrey[5],
-    },
-    [COMPLETED]: {
-        color: green[8],
-        backgroundColor: green[3],
-    },
-    [DEFERRED]: {
-        color: red[6],
-        backgroundColor: red[2],
-    },
-}
-
 module.exports = {
-    STATUS_SELECT_COLORS,
     colors,
     antGlobalVariables,
 }
