+++ apps/condo/domains/property/constants/property.example.js
