+++ apps/condo/domains/property/schema/Property.js
