+++ apps/condo/domains/property/schema/index.js
