+++ apps/condo/domains/ticket/constants/statusTypes.js
