+++ apps/condo/domains/ticket/schema/Ticket.test.js
@@ -1,7 +1,7 @@
 /**
  * Generated by `createschema ticket.Ticket organization:Text; statusReopenedCounter:Integer; statusReason?:Text; status:Relationship:TicketStatus:PROTECT; number?:Integer; client?:Relationship:User:SET_NULL; clientName:Text; clientEmail:Text; clientPhone:Text; operator:Relationship:User:SET_NULL; assignee?:Relationship:User:SET_NULL; classifier:Relationship:TicketClassifier:PROTECT; details:Text; meta?:Json;`
  */
-const { makeClientWithProperty } = require('@condo/domains/property/scheema/Property.test')
+const { makeClientWithProperty } = require('@condo/domains/property/schema/Property.test')
 const { NUMBER_RE, UUID_RE, DATETIME_RE, makeClient } = require('@core/keystone/test.utils')
 
 const { Ticket, createTestTicket, updateTestTicket } = require('@condo/domains/ticket/utils/testSchema')
