+++ apps/condo/index.js
@@ -37,7 +37,7 @@ registerSchemas(keystone, [
     require('@condo/domains/user/schema'),
     require('./schema/User'),
     require('./schema/Organization'),
-    require('@condo/domains/property/scheema'),
+    require('@condo/domains/property/schema'),
     require('@condo/domains/billing/schema'),
     require('@condo/domains/ticket/schema'),
 ])
